+++ samples/edd-sample-theme/functions.php
@@ -29,7 +29,7 @@ function edd_test_theme_updater() {
 		'api_url'     => 'http://eddstore.test', // Site where EDD is hosted.
 		'item_name'   => 'EDD Test Theme', // Name of theme.
 		'item_id'     => 27, // Item ID from Downloads page.
-		'theme_slug'  => 'edd-test-theme', // Theme slug.
+		'slug'        => 'edd-test-theme', // Theme slug.
 		'version'     => '1.0', // The current version of this theme.
 		'author'      => 'Andy Fragen', // The author of this theme.
 		'download_id' => '', // Optional, used for generating a license renewal link.