+++ src/License_Form.php
@@ -19,11 +19,11 @@ if ( ! defined( 'ABSPATH' ) ) {
  */
 class License_Form {
 	/**
-	 * Create generic license form table.
+	 * Create generic license form table row for each add-on.
 	 *
 	 * Can override using `edd_sl_license_form_table` filter.
 	 *
-	 * @param string $slug    Slug for EDD SL Add-on.
+	 * @param array  $addon    Array of add-on data.
 	 * @param string $license EDD SL license.
 	 * @param string $status  'valid' or 'invalid'.
 	 * @param string $message Activation/deactivation messsage.
@@ -31,49 +31,45 @@ class License_Form {
 	 *
 	 * @return void
 	 */
-	public function table( $slug, $license, $status, $message, $strings ) {
+	public function settings_row( $addon, $license, $status, $message, $strings ) {
+		$slug     = $addon['slug'];
+		$name     = $addon['item_name'];
+		$dashicon = 'plugin' === $addon['type'] ? '<span class="dashicons dashicons-admin-plugins"></span>&nbsp;&nbsp;' : '&nbsp;';
+		$dashicon = 'theme' === $addon['type'] ? '<span class="dashicons dashicons-admin-appearance"></span>&nbsp;&nbsp;' : $dashicon;
 		?>
-		<table class="form-table">
-				<tbody>
-					<tr valign="top">
-						<th scope="row" valign="top">
-							<?php esc_html_e( 'License Key', 'edd-sl-updater' ); ?>
-						</th>
-						<td>
-							<input id="<?php echo esc_attr( $slug ); ?>_license_key" name="<?php echo esc_attr( $slug ); ?>_license_key" type="text" class="regular-text" value="<?php echo esc_attr( $license, 'edd-sl-updater' ); ?>" />
-							<label class="description" for="<?php echo esc_attr( $slug ); ?>_license_key"></label>
-							<p class="description">
-								<?php echo esc_html( $message ); ?>
-							</p>
-						</td>
-					</tr>
-					<?php
-					if ( $license ) {
-						?>
-						<tr valign="top">
-							<th scope="row" valign="top">
-								<?php echo esc_attr( $strings['activate-license'] ); ?>
-							</th>
-							<td>
-								<?php
-								wp_nonce_field( $slug . '_nonce', $slug . '_nonce' );
-								if ( 'valid' === $status ) {
-									?>
-								<input type="submit" class="button-secondary" name="<?php echo esc_attr( $slug ); ?>_license_deactivate" value="<?php echo esc_attr( $strings['deactivate-license'] ); ?>"/>
-									<?php
-								} else {
-									?>
-								<input type="submit" class="button-secondary" name="<?php echo esc_attr( $slug ); ?>_license_activate" value="<?php echo esc_attr( $strings['activate-license'] ); ?>"/>
-									<?php
-								}
-								?>
-							</td>
-						</tr>
+			<tr valign="top">
+				<th scope="row" valign="top">
+					<?php echo( $dashicon . esc_html( $name ) ); ?>
+				</th>
+				<td>
+					<input id="<?php echo esc_attr( $slug ); ?>_license_key" name="<?php echo esc_attr( $slug ); ?>_license_key" type="text" class="regular-text" value="<?php echo esc_attr( $license, 'edd-sl-updater' ); ?>" />
+					<label class="description" for="<?php echo esc_attr( $slug ); ?>_license_key"></label>
+					<p class="description">
+						<?php echo esc_html( $message ); ?>
+					</p>
+				</td>
+			<?php
+			if ( $license ) {
+				?>
+					<td>
 						<?php
-					}
-					?>
-				</tbody>
-			</table>
+						wp_nonce_field( $slug . '_nonce', $slug . '_nonce' );
+						if ( 'valid' === $status ) {
+							?>
+						<input type="submit" class="button-secondary" name="<?php echo esc_attr( $slug ); ?>_license_deactivate" value="<?php echo esc_attr( $strings['deactivate-license'] ); ?>"/>
+							<?php
+						} else {
+							?>
+						<input type="submit" class="button-secondary" name="<?php echo esc_attr( $slug ); ?>_license_activate" value="<?php echo esc_attr( $strings['activate-license'] ); ?>"/>
+							<?php
+						}
+						?>
+					</td>
+
+				<?php
+			}
+			?>
+			</tr>
 		<?php
 	}
 }