+++ src/Theme_Updater_Admin.php
@@ -17,7 +17,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 /**
  * Class Theme_Updater_Admin
  */
-class Theme_Updater_Admin {
+class Theme_Updater_Admin extends Settings {
 	use API_Common;
 
 	/**
@@ -27,14 +27,16 @@ class Theme_Updater_Admin {
 	 * @var string
 	 */
 	protected $api_url     = null;
-	protected $theme_slug  = null;
+	protected $slug        = null;
 	protected $item_name   = null;
 	protected $item_id     = null;
+	protected $license     = null;
 	protected $version     = null;
 	protected $author      = null;
 	protected $download_id = null;
 	protected $renew_url   = null;
 	protected $strings     = null;
+	protected $theme_data  = null;
 
 	/**
 	 * Class constructor.
@@ -46,7 +48,7 @@ class Theme_Updater_Admin {
 			$config,
 			[
 				'api_url'     => 'http://easydigitaldownloads.com',
-				'theme_slug'  => get_template(),
+				'slug'        => get_template(),
 				'item_name'   => '',
 				'item_id'     => '',
 				'download_id' => '',
@@ -63,7 +65,8 @@ class Theme_Updater_Admin {
 		$this->item_name   = $config['item_name'];
 		$this->item_id     = $config['item_id'];
 		$this->download_id = $config['download_id'];
-		$this->theme_slug  = sanitize_key( $config['theme_slug'] );
+		$this->slug        = sanitize_key( $config['slug'] );
+		$this->license     = trim( get_option( $this->slug . '_license_key' ) );
 		$this->version     = $config['version'];
 		$this->author      = $config['author'];
 		$this->renew_url   = $config['renew_url'];
@@ -71,11 +74,12 @@ class Theme_Updater_Admin {
 
 		// Populate version fallback.
 		if ( empty( $config['version'] ) ) {
-			$theme         = wp_get_theme( $this->theme_slug );
+			$theme         = wp_get_theme( $this->slug );
 			$this->version = $theme->get( 'Version' );
 		}
 
-		$this->strings = $this->get_strings();
+		$this->strings    = $this->get_strings();
+		$this->theme_data = $config;
 
 		/**
 		 * Fires after the theme $config is setup.
@@ -94,12 +98,12 @@ class Theme_Updater_Admin {
 	 */
 	public function load_hooks() {
 		add_action( 'init', [ $this, 'updater' ] );
-		add_action( 'admin_menu', [ $this, 'license_menu' ] );
 		add_action( 'admin_init', [ $this, 'register_option' ] );
 		add_action( 'admin_init', [ $this, 'license_action' ] );
 		add_action( 'admin_notices', [ $this, 'show_error' ] );
-		add_action( 'update_option_' . $this->theme_slug . '_license_key', [ $this, 'activate_license' ], 10, 2 );
+		add_action( 'admin_init', [ $this, 'update_settings' ] );
 		add_filter( 'http_request_args', [ $this, 'disable_wporg_request' ], 5, 2 );
+		add_filter( 'edd_sl_updater_add_admin_page', [ $this, 'license_page' ] );
 	}
 
 	/**
@@ -111,7 +115,7 @@ class Theme_Updater_Admin {
 		}
 
 		// If there is no valid license key status, don't allow updates.
-		if ( 'valid' !== get_option( $this->theme_slug . '_license_key_status', false ) ) {
+		if ( 'valid' !== get_option( $this->slug . '_license_key_status', false ) ) {
 			return;
 		}
 
@@ -119,7 +123,7 @@ class Theme_Updater_Admin {
 			[
 				'api_url'   => $this->api_url,
 				'version'   => $this->version,
-				'license'   => trim( get_option( $this->theme_slug . '_license_key' ) ),
+				'license'   => trim( get_option( $this->slug . '_license_key' ) ),
 				'item_name' => $this->item_name,
 				'item_id'   => $this->item_id,
 				'author'    => $this->author,
@@ -130,101 +134,57 @@ class Theme_Updater_Admin {
 	}
 
 	/**
-	 * Adds a menu item for the theme license under the appearance menu.
-	 */
-	public function license_menu() {
-		$strings = $this->strings;
-
-		add_theme_page(
-			$strings['theme-license'],
-			$strings['theme-license'],
-			'manage_options',
-			$this->theme_slug . '-license',
-			[ $this, 'license_page' ]
-		);
-	}
-
-	/**
 	 * Registers the option used to store the license key in the options table.
 	 */
 	public function register_option() {
 		register_setting(
-			$this->theme_slug . '-license',
-			$this->theme_slug . '_license_key',
-			[ $this, 'sanitize_license' ]
+			$this->slug . '-license',
+			$this->slug . '_license_key',
+			[]
 		);
-	}
-
-	/**
-	 * Sanitizes the license key.
-	 *
-	 * @param  string $new License key that was submitted.
-	 * @return string $new Sanitized license key.
-	 */
-	public function sanitize_license( $new ) {
-		$old = get_option( $this->theme_slug . '_license_key' );
-
-		if ( $old && $old !== $new ) {
-			// New license has been entered, so must reactivate
-			delete_option( $this->theme_slug . '_license_key_status' );
-			delete_transient( $this->theme_slug . '_license_message' );
-		}
-
-		return $new;
 	}
 
 	/**
 	 * Outputs the markup used on the theme license page.
 	 */
 	public function license_page() {
-		$license = trim( get_option( $this->theme_slug . '_license_key' ) );
-		$status  = get_option( $this->theme_slug . '_license_key_status', false );
+		$license_check = $this->check_license( $this->slug );
+		$license       = trim( get_option( $this->slug . '_license_key' ) );
+		$status        = $license_check['license'];
 
 		// Checks license status to display under license key.
 		if ( ! $license ) {
 			$message = $strings['enter-key'];
 		} else {
-			// delete_transient( $this->theme_slug . '_license_message' );
-			if ( ! get_transient( $this->theme_slug . '_license_message', false ) ) {
-				set_transient( $this->theme_slug . '_license_message', $this->check_license( $this->theme_slug ), ( 60 * 60 * 24 ) );
+			delete_transient( $this->slug . '_license_message' );
+			if ( ! get_transient( $this->slug . '_license_message', false ) ) {
+				set_transient( $this->slug . '_license_message', $license_check['message'], ( 60 * 60 * 24 ) );
 			}
-			$message = get_transient( $this->theme_slug . '_license_message' );
-		} ?>
-		<div class="wrap">
-			<h2>
-			<?php echo esc_attr( $this->strings['theme-license'] . ' - ' . $this->item_name ); ?>
-			</h2>
-			<form method="post" action="options.php">
-				<?php
-				settings_fields( $this->theme_slug . '-license' );
+			$message = get_transient( $this->slug . '_license_message' );
+		}
+		settings_fields( $this->slug . '-license' );
+		$form_table_row = ( new License_Form() )->settings_row( $this->theme_data, $license, $status, $message, $this->strings );
 
-				$form_table = ( new License_Form() )->table( $this->theme_slug, $license, $status, $message, $this->strings );
-
-				/**
-				 * Filter to echo a customized license form table.
-				 *
-				 * @since 1.0.0
-				 *
-				 * @param string $form_table Table HTML for a license page setting.
-				 * @param string $slug       EDD SL Add-on slug.
-				 * @param string $license    EDD SL license.
-				 * @param string $status     License status.
-				 * @param string $message    License message.
-				 * @param array  $strings    Messaging strings.
-				 */
-				echo apply_filters( 'edd_sl_license_form_table', $form_table, $this->theme_slug, $license, $status, $message, $this->strings );
-
-				submit_button();
-				?>
-			</form>
-		<?php
+		/**
+		 * Filter to echo a customized license form table.
+		 *
+		 * @since 1.0.0
+		 *
+		 * @param string $form_table Table HTML for a license page setting.
+		 * @param string $slug       EDD SL Add-on slug.
+		 * @param string $license    EDD SL license.
+		 * @param string $status     License status.
+		 * @param string $message    License message.
+		 * @param array  $strings    Messaging strings.
+		 */
+		echo apply_filters( 'edd_sl_license_form_table', $form_table_row, $this->theme_data, $license, $status, $message, $this->strings );
 	}
 
 	/**
 	 * Activates the license key.
 	 */
 	public function activate_license() {
-		$license = trim( get_option( $this->theme_slug . '_license_key' ) );
+		$license = trim( get_option( $this->slug . '_license_key' ) );
 
 		// Data to send in our API request.
 		$api_params = [
@@ -237,7 +197,7 @@ class Theme_Updater_Admin {
 
 		$license_data = $this->get_api_response( $this->api_url, $api_params );
 
-		if ( $license_data->success && isset( $license_data->error ) ) {
+		if ( ! $license_data->success && isset( $license_data->error ) ) {
 			switch ( $license_data->error ) {
 				case 'expired':
 					$message = sprintf(
@@ -268,10 +228,9 @@ class Theme_Updater_Admin {
 			}
 		}
 
-		// $response->license will be either "active" or "inactive".
-		if ( $license_data && isset( $license_data->license ) ) {
-			update_option( $this->theme_slug . '_license_key_status', $license_data->license );
-			delete_transient( $this->theme_slug . '_license_message' );
+		if ( isset( $license_data, $license_data->license ) ) {
+			update_option( $this->slug . '_license_key_status', $license_data->license );
+			delete_transient( $this->slug . '_license_message' );
 		}
 
 		if ( ! empty( $message ) ) {
@@ -289,7 +248,7 @@ class Theme_Updater_Admin {
 	 */
 	public function deactivate_license() {
 		// Retrieve the license from the database.
-		$license = trim( get_option( $this->theme_slug . '_license_key' ) );
+		$license = trim( get_option( $this->slug . '_license_key' ) );
 
 		// Data to send in our API request.
 		$api_params = [
@@ -316,8 +275,8 @@ class Theme_Updater_Admin {
 
 		// $license_data->license will be either "deactivated" or "failed".
 		if ( $license_data && ( 'deactivated' === $license_data->license ) ) {
-			delete_option( $this->theme_slug . '_license_key_status' );
-			delete_transient( $this->theme_slug . '_license_message' );
+			delete_option( $this->slug . '_license_key_status' );
+			delete_transient( $this->slug . '_license_message' );
 		}
 		$this->redirect( $error_data );
 	}
@@ -334,7 +293,7 @@ class Theme_Updater_Admin {
 		}
 
 		// If download_id was passed in the config, a renewal link can be constructed.
-		$license_key = trim( get_option( $this->theme_slug . '_license_key', false ) );
+		$license_key = trim( get_option( $this->slug . '_license_key', false ) );
 		if ( ! empty( $this->download_id ) && $license_key ) {
 			$url  = esc_url( $this->api_url );
 			$url .= '/checkout/?edd_license_key=' . $license_key . '&download_id=' . $this->download_id;
@@ -352,14 +311,14 @@ class Theme_Updater_Admin {
 	 * @since 1.0.0
 	 */
 	public function license_action() {
-		if ( isset( $_POST[ $this->theme_slug . '_license_activate' ] ) ) {
-			if ( check_admin_referer( $this->theme_slug . '_nonce', $this->theme_slug . '_nonce' ) ) {
+		if ( isset( $_POST[ $this->slug . '_license_activate' ] ) ) {
+			if ( check_admin_referer( $this->slug . '_nonce', $this->slug . '_nonce' ) ) {
 				$this->activate_license();
 			}
 		}
 
-		if ( isset( $_POST[ $this->theme_slug . '_license_deactivate' ] ) ) {
-			if ( check_admin_referer( $this->theme_slug . '_nonce', $this->theme_slug . '_nonce' ) ) {
+		if ( isset( $_POST[ $this->slug . '_license_deactivate' ] ) ) {
+			if ( check_admin_referer( $this->slug . '_nonce', $this->slug . '_nonce' ) ) {
 				$this->deactivate_license();
 			}
 		}