+++ CHANGES.md
@@ -9,3 +9,4 @@
 * created `class Init` with universal instantiator function, just add configuration array
 * created filter hook `edd_sl_license_form_table` to easily customize the license form
 * switch to `site_transient_update_{plugins|themes}` filter
+* created Settings page for licenses instead of appending to menus